add authorization

+++ app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Services\ProductService;
 use App\Http\Resources\ProductResource;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
+use App\Models\Product;
 use Exception;
 use Illuminate\Support\Facades\Log;
 
@@ -21,7 +22,6 @@ class ProductController extends Controller
     // Lấy tất cả sản phẩm
     public function index()
     {
-
         return ProductResource::collection(
             $this->productService->getAll()->load('category')
         );
@@ -41,7 +41,7 @@ class ProductController extends Controller
     // Tạo sản phẩm mới
     public function store(StoreProductRequest $request)
     {
-
+        $this->authorize('create', Product::class);
         $product = $this->productService->create($request->validated());
         return new ProductResource($product);
     }
@@ -49,6 +49,9 @@ class ProductController extends Controller
     public function update(UpdateProductRequest $request, $id)
     {
         try {
+            $product = Product::findOrFail($id);
+            $this->authorize('update', $product);
+            // Cập nhật sản phẩm
             $updatedProduct = $this->productService->update($id, $request->validated());
             return response()->json([
                 'product' => new ProductResource($updatedProduct),
@@ -59,13 +62,16 @@ class ProductController extends Controller
             return response()->json(['message' => 'Product not found'], 404);
         }
     }
+
     // Xóa sản phẩm
     public function destroy($id)
     {
         try {
+            $this->authorize('delete', Product::class);
             $this->productService->delete($id);
-            return response()->json(['message' => 'Product deleted successfully'], 204);
+            return response()->json(['message' => 'Product deleted successfully'], 202);
         } catch (Exception $e) {
+            Log::error($e->getMessage());
             return response()->json(['message' => 'Product not found'], 404);
         }
     }
